requirements.txt
astroid==2.4.0
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
future==0.18.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pkg-resources==0.0.0
praw==7.0.0
prawcore==1.3.0
pycodestyle==2.5.0
pylint==2.5.0
python-dotenv==0.13.0
python-twitter==3.5
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
wrapt==1.12.1
